clean up internal and isolate manifest
